Refactor: Extracted RabbitMq server controller from IntegrationFixture
Simplifying IntegrationFixture.
Make reuse of code for controlling RabbitMQ server simpler.
